requirements.txt
flask
pillow
numpy
opencv-python
azure-cognitiveservices-vision-computervision
python-dotenv
pandas
openpyxl
